Overhaul the main program to prevent deadlocking
The previous naive approach relied on setting up and loading the seccomp
filter in the first process, then immediately executing the desired
process. This had the unfortunate effect of causing deadlocks in the
parent when part of its syscall emulation involved a syscall that was
also part of the child's seccomp filter (after all, the filter was also
active in the parent).

Now we have a multi-stage approach:

* The first process (stage 1) sets up a socketpair and produces a child
* The child (stage 2) sets up the filter and sends the notification FD
  and PID of the end-user's process back to the stage 1 process, then
  exits.
* Stage 2 is now over. Stage 1 can receive all seccomp notifications
  from the third process and its descendants without deadlocking if it
  itself calls those same syscalls.

Part of the magic is taking advantage of SCM_RIGHTS to pass the
notification FD around as well as PR_SET_CHILD_SUBREAPER to make sure
none of the descendants escape. The latter will also be useful in the
future when it is necessary to track the uid, euid, et al of the
children.